requirements.txt
streamlit
langchain
langchain-groq
langchain-community
pandas
openpyxl
python-pptx
pypdf
pillow
fpdf
python-dotenv
tiktoken
pytesseract
PyPDF2
langchain-mcp-adapters
mcp
langgraph
praw
requests
wikipedia
google-api-python-client
transformers
mcp_use